refactor: インターフェース削除
